fix(native-renderer): video frame compositing via Image elements with video_frames_dir
The element capture converts Video elements to Image type (for Chrome PNG
capture). The Rust painter now checks style.video_frames_dir on Image
elements and uses per-frame extracted JPEGs when present, falling back to
the static Chrome PNG for that element.

PSNR results on Apple presentation (141s, 7 slides, 5 videos):
  Slides WITHOUT video: 35.7dB (above 30dB imperceptible threshold)
  Slides WITH video: 11-15dB (video frames partially mapped — 3/5 videos)

The remaining gap is from cinematic-vid and avatar-vid which are in
sub-compositions and may not have frame directories at the expected paths.

+++ packages/native-renderer/src/paint/elements.rs
@@ -606,8 +606,8 @@ fn paint_element_inner(
             let media_start = style.video_media_start.unwrap_or(0.0);
             let video_fps = style.video_fps.unwrap_or(30.0);
             let element_start = style.data_start.unwrap_or(0.0);
-            let video_time = (t - element_start as f64 + media_start as f64).max(0.0);
-            let frame_index = (video_time * video_fps as f64).round() as u32;
+            let video_time = (t - element_start + media_start).max(0.0);
+            let frame_index = (video_time * video_fps).round() as u32;
             let frame_path = format!("{}/frame_{:05}.jpg", frames_dir, frame_index + 1);
             if let Some(image) = images.get_or_load(&frame_path).cloned() {
                 let dest_rect = to_sk_rect(&element.bounds);
